chore(connect): move pathUtils from connect to crypto-utils
