+++ packages/connect/package.json
@@ -51,6 +51,7 @@
     "dependencies": {
         "@trezor/blockchain-link": "workspace:^2.1.8",
         "@trezor/connect-common": "workspace:0.0.12",
+        "@trezor/crypto-utils": "workspace:1.0.0",
         "@trezor/transport": "workspace:^1.1.8",
         "@trezor/utils": "workspace:^9.0.6",
         "@trezor/utxo-lib": "workspace:^1.0.4",
