+++ packages/crypto-utils/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+    rules: {
+        'no-bitwise': 'off',
+    },
+};
