+++ packages/crypto-utils/package.json
@@ -4,9 +4,14 @@
     "private": true,
     "license": "See LICENSE.md in repo root",
     "sideEffects": false,
-    "main": "src/index",
+    "main": "lib/index",
     "scripts": {
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
-        "type-check": "tsc --build"
+        "type-check": "tsc --build",
+        "build:lib": "rimraf ./lib && yarn tsc --build tsconfig.lib.json"
+    },
+    "devDependencies": {
+        "rimraf": "^4.1.2",
+        "typescript": "4.9.5"
     }
 }
