+++ packages/crypto-utils/src/index.ts
@@ -1,1 +1,2 @@
 export { bip39 } from './bip39';
+export * from './path';
