+++ packages/crypto-utils/tsconfig.lib.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../../tsconfig.lib.json",
+    "compilerOptions": {
+        "outDir": "lib"
+    },
+    "include": ["./src"]
+}
